+++ src/lib/Scenes/Collection/Collection.tsx
@@ -81,10 +81,11 @@ export class Collection extends Component<CollectionProps, CollectionState> {
 
 export const CollectionContainer = createFragmentContainer(Collection, {
   collection: graphql`
-    fragment Collection_collection on MarketingCollection {
+    fragment Collection_collection on MarketingCollection
+      @argumentDefinitions(screenWidth: { type: "Int", defaultValue: 500 }) {
       ...CollectionHeader_collection
       ...CollectionArtworks_collection
-      ...FeaturedArtists_collection
+      ...FeaturedArtists_collection @arguments(screenWidth: $screenWidth)
     }
   `,
 })